Deleted
Archivo cargarCadena borrado y su funcion se movio a utilidades.cpp
