+++ utilidades.h
@@ -3,6 +3,7 @@
 
 #include <string>
 
+void cargarCadena(char *, int);
 void toUpperCase(char *text);
 void centrar_texto(string palabra, char dato, size_t cantidad);
